textract_probe/requirements.txt
boto3>=1.35.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
pytest>=8.0.0
